Use the new workloadClusterLabel in the namespace reconciler.
This commit adds support to the namespace controller to use the new WorkloadClusterLabel `cluster.internal.workloads.kcp.dev/<WorkloadClusterName>` and replace the previously used label `workloads.kcp.dev/cluster`

To reduce the required changes and make integrating with the syncer virtual workspace easier, we introduce another label that is used to keep backward compatibility with the current code and e2e tests: `workloads.kcp.dev/scheduled-cluster`. This label is internal, and it's used only to specify the scheduled cluster on a namespace in non-advanced-scheduling mode.